+++ src/tc-renderer/lib/startup/angular-stuff.js
@@ -1,17 +1,17 @@
 import '../../ng/app'
 
-import '../../ng/elements/login/login'
-import '../../ng/elements/settings-panel/settings-panel'
-import '../../ng/elements/side-toolbar/side-toolbar'
-import '../../ng/elements/chat-input/chat-input'
-import '../../ng/elements/chat-output/chat-output'
-import '../../ng/elements/chat-tabs/chat-tabs'
-import '../../ng/elements/user-panel/user-panel'
-import '../../ng/elements/thumbnail/thumbnail'
-import '../../ng/elements/chatters/chatters'
-import '../../ng/elements/chatters-tabs/chatters-tabs'
-import '../../ng/elements/add-channel/add-channel'
-import '../../ng/elements/emote-menu/emote-menu'
+import '../../ng/components/login/login'
+import '../../ng/components/settings-panel/settings-panel'
+import '../../ng/components/side-toolbar/side-toolbar'
+import '../../ng/components/chat-input/chat-input'
+import '../../ng/components/chat-output/chat-output'
+import '../../ng/components/chat-tabs/chat-tabs'
+import '../../ng/components/user-panel/user-panel'
+import '../../ng/components/thumbnail/thumbnail'
+import '../../ng/components/chatters/chatters'
+import '../../ng/components/chatters-tabs/chatters-tabs'
+import '../../ng/components/add-channel/add-channel'
+import '../../ng/components/emote-menu/emote-menu'
 
 import '../../ng/providers/store'
 import '../../ng/providers/settings'
