+++ src/tc-renderer/ng/components/add-channel/add-channel.js
