+++ src/tc-renderer/ng/components/chat-input/chat-input.styl
