+++ src/tc-renderer/ng/components/chat-output/chat-output.js
