+++ src/tc-renderer/ng/components/chat-tabs/chat-tabs.js
