+++ src/tc-renderer/ng/components/chatters-tabs/chatters-tabs.styl
