+++ src/tc-renderer/ng/components/chatters/chatters.js
