+++ src/tc-renderer/ng/components/emote-menu/emote-menu.js
