+++ src/tc-renderer/ng/components/settings-panel/settings-gui-dialog.pug
