+++ src/tc-renderer/ng/components/settings-panel/settings-panel.pug
