+++ src/tc-renderer/ng/components/side-toolbar/side-toolbar.styl
