+++ src/tc-renderer/ng/components/thumbnail/thumbnail.pug
