+++ src/tc-renderer/ng/components/thumbnail/thumbnail.styl
