+++ src/tc-renderer/ng/components/user-panel/user-panel.styl
